refactor: 불필요한 console.log 문 제거
