+++ src/apis/ranking/index.ts
@@ -2,17 +2,17 @@ import type { RankingFilterOption } from '@/types';
 import type { Product } from '@/types/product';
 
 import instance from '../instance';
-import API from '../path';
+import API from '../path.constants';
 
 export const getRanking = async ({
   rankType,
   targetType,
 }: RankingFilterOption): Promise<Product[]> => {
-  const res = await instance.get(API.RANKING, {
+  const { data } = await instance.get(API.RANKING, {
     params: {
       rankType,
       targetType,
     },
   });
-  return res.data!.products;
+  return data.products;
 };
