+++ src/apis/theme/index.ts
@@ -1,9 +1,9 @@
 import type { Theme } from '@/types/product';
 
 import instance from '../instance';
-import API from '../path';
+import API from '../path.constants';
 
 export const getThemes = async (): Promise<Theme[]> => {
-  const res = await instance.get(API.THEMES);
-  return res.data!.themes;
+  const { data } = await instance.get(API.THEMES);
+  return data.themes;
 };
